refactor: Refactor @EmbbededId TopPercentageHistoryId as @IdClass TopPercenrtaggeHistoryId
