+++ family/src/main/kotlin/com/bibbi/domain/model/TopPercentageHistory.kt
@@ -3,9 +3,13 @@ package com.bibbi.domain.model
 import java.time.LocalDateTime
 
 data class TopPercentageHistory(
-    val topPercentageHistoryId: TopPercentageHistoryId,
+    val familyId: String,
+    val year: Int,
+    val month: Int,
+
     val family: Family,
     var topPercentage: Int,
+
     val createdAt : LocalDateTime?,
     val updatedAt : LocalDateTime?,
 )
