--- family/src/main/kotlin/com/bibbi/domain/model/TopPercentageHistoryId.kt
@@ -1,7 +1,0 @@
-package com.bibbi.domain.model
-
-data class TopPercentageHistoryId(
-    val familyId: String,
-    val year: Int,
-    val month: Int,
-)
